\name{Bering-package}
\alias{Bering-package}
\alias{Bering}
\docType{package}
\title{
\packageTitle{Bering}
}
\description{
\packageDescription{Bering}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bering}
\packageIndices{Bering}
Use the functions in this package to conduct randomization-based hypothesis testing for climate data. In particular, this package is useful for comparing the mean values during laminated intervals and the mean values during non-laminated intervals.
}
\author{
\packageAuthor{Bering}

Maintainer: \packageMaintainer{Bering}
}


\keyword{ package }
\keyword{ time-series}
\keyword{climate}


